+++ Character.h
@@ -49,7 +49,7 @@ class Character {
     float getVy();
     void setPosition(float x, float y);
 
-    void init(uint16_t spawnX, uint16_t spawnY);
+    void init(uint8_t spawnX, uint8_t spawnY);
     
 };
 #endif
